refactor: move `interpolate` function to Extensions
